+++ SeriesGuide/src/main/java/com/battlelancer/seriesguide/provider/SeriesGuideContract.java
@@ -745,6 +745,8 @@ public class SeriesGuideContract {
 
         public static final String SELECTION_WATCHLIST = Movies.IN_WATCHLIST + "=1";
 
+        public static final String SELECTION_WATCHED = Movies.WATCHED + "=1";
+
         /** Default sort order. */
         public static final String SORT_TITLE_ALPHABETICAL = Movies.TITLE + " COLLATE NOCASE ASC";