+++ SeriesGuide/src/main/java/com/battlelancer/seriesguide/ui/TvdbAddFragment.java
@@ -16,9 +16,9 @@
 
 package com.battlelancer.seriesguide.ui;
 
-import android.content.Context;
-import android.os.AsyncTask;
 import android.os.Bundle;
+import android.support.v4.app.LoaderManager;
+import android.support.v4.content.Loader;
 import android.view.KeyEvent;
 import android.view.LayoutInflater;
 import android.view.Menu;
@@ -37,16 +37,10 @@ import butterknife.ButterKnife;
 import butterknife.InjectView;
 import com.battlelancer.seriesguide.R;
 import com.battlelancer.seriesguide.items.SearchResult;
+import com.battlelancer.seriesguide.loaders.TvdbAddLoader;
 import com.battlelancer.seriesguide.settings.SearchSettings;
-import com.battlelancer.seriesguide.thetvdbapi.TheTVDB;
-import com.battlelancer.seriesguide.thetvdbapi.TvdbException;
 import com.battlelancer.seriesguide.util.SearchHistory;
-import com.uwetrottmann.androidutils.AndroidUtils;
-import de.greenrobot.event.EventBus;
 import java.util.ArrayList;
-import java.util.LinkedList;
-import java.util.List;
-import timber.log.Timber;
 
 public class TvdbAddFragment extends AddFragment {
 
@@ -54,20 +48,11 @@ public class TvdbAddFragment extends AddFragment {
         return new TvdbAddFragment();
     }
 
-    public static class TvdbAddResultsEvent {
-        public int emptyTextResId;
-        public List<SearchResult> results;
-
-        public TvdbAddResultsEvent(int emptyTextResId, List<SearchResult> results) {
-            this.emptyTextResId = emptyTextResId;
-            this.results = results;
-        }
-    }
+    private static final String KEY_QUERY = "search-query";
 
     @InjectView(R.id.buttonAddTvdbClear) ImageButton clearButton;
     @InjectView(R.id.editTextAddTvdbSearch) AutoCompleteTextView searchBox;
 
-    private SearchTask searchTask;
     private SearchHistory searchHistory;
     private ArrayAdapter<String> searchHistoryAdapter;
 
@@ -110,8 +95,11 @@ public class TvdbAddFragment extends AddFragment {
         // set in code as XML is overridden
         searchBox.setImeOptions(EditorInfo.IME_ACTION_SEARCH);
         searchBox.setInputType(EditorInfo.TYPE_CLASS_TEXT);
+        // drop-down is auto-shown on config change, ensure it is hidden when recreating views
+        searchBox.dismissDropDown();
 
-        setProgressVisible(false, false);
+        // set initial view states
+        setProgressVisible(true, false);
 
         return v;
     }
@@ -120,39 +108,22 @@ public class TvdbAddFragment extends AddFragment {
     public void onActivityCreated(Bundle savedInstanceState) {
         super.onActivityCreated(savedInstanceState);
 
-        // create an empty adapter to avoid displaying a progress indicator
-        if (adapter == null) {
-            adapter = new AddAdapter(getActivity(), R.layout.item_addshow,
-                    new ArrayList<SearchResult>());
-        }
-
-        if (!AndroidUtils.isNetworkConnected(getActivity())) {
-            setEmptyMessage(R.string.offline);
-        }
+        // create adapter
+        adapter = new AddAdapter(getActivity(), R.layout.item_addshow,
+                new ArrayList<SearchResult>());
 
         // setup search history
-        if (searchHistory == null || searchHistoryAdapter == null) {
-            searchHistory = new SearchHistory(getActivity(), SearchSettings.KEY_SUFFIX_THETVDB);
-            searchHistoryAdapter = new ArrayAdapter<>(getActivity(),
-                    android.R.layout.simple_dropdown_item_1line, searchHistory.getSearchHistory());
-            searchBox.setAdapter(searchHistoryAdapter);
-        }
+        searchHistory = new SearchHistory(getActivity(), SearchSettings.KEY_SUFFIX_THETVDB);
+        searchHistoryAdapter = new ArrayAdapter<>(getActivity(),
+                android.R.layout.simple_dropdown_item_1line, searchHistory.getSearchHistory());
+        searchBox.setAdapter(searchHistoryAdapter);
 
+        // load data
+        getLoaderManager().initLoader(AddActivity.AddPagerAdapter.SEARCH_TAB_DEFAULT_POSITION, null,
+                mTvdbAddCallbacks);
+
+        // enable menu
         setHasOptionsMenu(true);
-    }
-
-    @Override
-    public void onStart() {
-        super.onStart();
-
-        EventBus.getDefault().register(this);
-    }
-
-    @Override
-    public void onStop() {
-        super.onStop();
-
-        EventBus.getDefault().unregister(this);
     }
 
     @Override
@@ -181,78 +152,47 @@ public class TvdbAddFragment extends AddFragment {
     private void search() {
         searchBox.dismissDropDown();
 
-        // clear current search results
-        setSearchResults(new LinkedList<SearchResult>());
-
-        // nag about no connectivity
-        if (!AndroidUtils.isNetworkConnected(getActivity())) {
-            setEmptyMessage(R.string.offline);
-            return;
-        }
-
+        // extract query
         String query = searchBox.getText().toString().trim();
-        if (query.length() == 0) {
-            return;
-        }
-        // query for results
-        if (searchTask == null || searchTask.getStatus() == AsyncTask.Status.FINISHED) {
-            setProgressVisible(true, false);
-            searchTask = new SearchTask(getActivity());
-            AndroidUtils.executeOnPool(searchTask, query);
-        }
+
+        // do search
+        Bundle args = new Bundle();
+        args.putString(KEY_QUERY, query);
+        getLoaderManager().restartLoader(AddActivity.AddPagerAdapter.SEARCH_TAB_DEFAULT_POSITION,
+                args, mTvdbAddCallbacks);
+
         // update history
-        if (searchHistory.saveRecentSearch(query)) {
-            searchHistoryAdapter.clear();
-            searchHistoryAdapter.addAll(searchHistory.getSearchHistory());
+        if (query.length() > 0) {
+            if (searchHistory.saveRecentSearch(query)) {
+                searchHistoryAdapter.clear();
+                searchHistoryAdapter.addAll(searchHistory.getSearchHistory());
+            }
         }
     }
 
-    public void onEventMainThread(TvdbAddResultsEvent event) {
-        setEmptyMessage(event.emptyTextResId);
-        setSearchResults(event.results);
-        setProgressVisible(false, true);
-    }
+    private LoaderManager.LoaderCallbacks<TvdbAddLoader.Result> mTvdbAddCallbacks
+            = new LoaderManager.LoaderCallbacks<TvdbAddLoader.Result>() {
+        @Override
+        public Loader<TvdbAddLoader.Result> onCreateLoader(int id, Bundle args) {
+            setProgressVisible(true, false);
 
-    public static class SearchTask extends AsyncTask<String, Void, List<SearchResult>> {
-
-        private Context mContext;
-
-        public SearchTask(Context context) {
-            mContext = context;
+            String query = null;
+            if (args != null) {
+                query = args.getString(KEY_QUERY);
+            }
+            return new TvdbAddLoader(getActivity(), query);
         }
 
         @Override
-        protected List<SearchResult> doInBackground(String... params) {
-            List<SearchResult> results;
-
-            String query = params[0];
-
-            try {
-                results = TheTVDB.searchShow(mContext, query, false);
-                if (results.size() == 0) {
-                    // query again, but allow all languages
-                    results = TheTVDB.searchShow(mContext, query, true);
-                }
-            } catch (TvdbException e) {
-                Timber.e(e, "Searching show failed");
-                return null;
-            }
-
-            return results;
+        public void onLoadFinished(Loader<TvdbAddLoader.Result> loader, TvdbAddLoader.Result data) {
+            setSearchResults(data.results);
+            setEmptyMessage(data.emptyTextResId);
+            setProgressVisible(false, true);
         }
 
         @Override
-        protected void onPostExecute(List<SearchResult> result) {
-            TvdbAddResultsEvent event;
-            if (result == null) {
-                // display error in empty view
-                event = new TvdbAddResultsEvent(R.string.search_error,
-                        new LinkedList<SearchResult>());
-            } else {
-                // empty or there are shows
-                event = new TvdbAddResultsEvent(R.string.no_results, result);
-            }
-            EventBus.getDefault().post(event);
+        public void onLoaderReset(Loader<TvdbAddLoader.Result> loader) {
+            // keep existing data
         }
-    }
+    };
 }